Fixed the cause for "BUG: sleeping function called from invalid context at mm/slub.c" warning message.
In the functions ed_create and td_create, the memory is being
allocated using GFP_KERNEL flag. Since ed_create and td_create
functions are called in a context with spinlock held, the
GFP_ATOMIC flag will have to be used for memory allocation.
This message was observed when in USB OTG Host mass storage
test during connect, disconnect and data transfer.