S5PC110: TVOUT: Fixed issue - when mixer set to Interlaced mode, vertical scale up error
When mixer set to interlaced mode, and graphic layer is scaled up,
pixel data was broken. In that case, mixer scale up ftn. must not use
and intead of that ftn. change the mixer input source mode from interlaced
to progressive - at the result, i to p is same with vertical scale up(X2)