S5PC110: MFC: add consumed byte information in the decoder
In the H264/DivX/Xvid, some parser generates 2 samples in
the one time. Therefore, stream buffer has 2 packed data.
For example, PackedPB frame exists in the MPEG4, or two
fields can be in the stream buffer in case of H264.
That's why consumed byte information is needed right after
one data is decoded.
Signed-off-by: Jaeryul Oh <jaeryul.oh@samsung.com>