S5PC110: SYSTEM: Modify Settings to avoid I2S/PCM port conflict
In c110 machine code for initalize, all PCM and I2S GPIO port register
set to use for them. But in EVT1, code that set all GPIO port occurs a
conflict when I2S and PCM selected to use at once. Because of I2S0 and
PCM0 are share same GPIO ports. So, fixed it to two types of port set
selection in kernel config.

Signed-off-by: Seungwhan Youn <sw.youn@samsung.com>