S5PC110: ASoC: Modify audio driver to use XUSBXTI for codec's XTI
Modify codec's XTI resource to XUSBXTI from EPLL_OUT by XCLK_OUT. Because
EPLL is not guarantee a constant freq. rate. So it is hard to calcurate a
correct value to operate codec. XUSBXTI is always on 24MHz and divided it
by 2, it is same as privious X1(12MHz) freq. So, changed it.

Signed-off-by: Seungwhan Youn <sw.youn@samsung.com>