ASoC: Allow mulitple usage of codec_dai
In a situation when we have two cpu_dai's that ultimately play audio through
the same codec, we need some mechanism to avoid shutting down the codec when
one i/f is close while the other is still active.

For that, we change semantics of the snd_soc_dai.active flag from indicator
to reference counter. The codec_dai is shutdown only when it's counter
reaches zero.

Signed-off-by: Jassi Brar <jassi.brar@samsung.com>