S5P: ASoC: Remove check I2S active condition for normal sleep
On normal sleep mode of power management, when system goes to sleep,
AP cut off the I2S power. So, all values of I2S control registers are
erased. Because of this reason, I2S lost it's pre-set values and this
makes I2S not working after resume. So, I2S is active or not, all re-
gisters are to be back-up to keep it's pre-set values.

Signed-off-by: Seungwhan Youn <sw.youn@samsung.com>