S5PC110: DVFS: SCLKA2M clock is returned to 200Mhz configuration
When changing APLL, MPLL is used as a temporary clock source of
MFC and G3D. MFC and G3D's default clock source is SCLKA2M. This clock
is divided by DIVA2M from APLL output. The maximum clock which can be
generated by controlling DIVA2M is 400Mhz. This restriction is condsidered
in this fix.

Signed-off-by: Jaecheol Lee <jc.lee@samsung.com>